+++ SimpleBank/SimpleBank.Core/Commands/CreditByAccountId.cs
@@ -1,0 +1,6 @@
+﻿namespace SimpleBank.Core.Commands
+{
+    public class CreditByAccountId : OperationCommand
+    {
+    }
+}
