+++ SimpleBank/SimpleBank.Core/Commands/DebitByAccountId.cs
@@ -1,0 +1,6 @@
+﻿namespace SimpleBank.Core.Commands
+{
+    public class DebitByAccountId : OperationCommand
+    {
+    }
+}
