+++ SimpleBank/SimpleBank.Core/Models/OperationType.cs
@@ -1,0 +1,8 @@
+﻿namespace SimpleBank.Core.Models
+{
+    public enum OperationType : byte
+    {
+        Debit = 1,
+        Credit = 2,
+    }
+}
