+++ SimpleBank/SimpleBank.Core/Services/IOperationService.cs
@@ -1,0 +1,17 @@
+﻿using SimpleBank.Core.Commands;
+
+namespace SimpleBank.Core.Services
+{
+    public interface IOperationService
+    {
+        /// <summary>
+        /// ანგარიშიდან თანხის ჩამოჭრის ოპერაცია
+        /// </summary>
+        long Debit(DebitByAccountId command);
+
+        /// <summary>
+        /// ანგარიშზე თანხის ჩარიხვის ოპერაცია
+        /// </summary>
+        void Credit(CreditByAccountId command);
+    }
+}
