+++ SimpleBank/SimpleBank.Core/Utils/DateTimeProvider.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace SimpleBank.Core.Utils
+{
+    public class DateTimeProvider : IDateTimeProvider
+    {
+        public DateTime Now => DateTime.Now;
+        public DateTime UtcNow => DateTime.UtcNow;
+    }
+}
