+++ SimpleBank/SimpleBank.Core/Utils/IDateTimeProvider.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace SimpleBank.Core.Utils
+{
+    public interface IDateTimeProvider
+    {
+        public DateTime Now { get; }
+        public DateTime UtcNow { get; }
+    }
+}
